Repository was moved from Controllers to Repositories directory
